+++ src/Sekiban.Core.DotNet/Dependency/SekibanEventSourcingDependency.cs
@@ -16,124 +16,7 @@ namespace Sekiban.Core.Dependency;
 public static class SekibanEventSourcingDependency
 {
     public static Assembly GetAssembly() => Assembly.GetExecutingAssembly();
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="builder"></param>
-    /// <param name="dependencyDefinition"></param>
-    /// <returns></returns>
-    public static IHostApplicationBuilder AddSekibanWithDependency(
-        this IHostApplicationBuilder builder,
-        IDependencyDefinition dependencyDefinition)
-    {
-        builder.Services.AddSekibanWithDependency(dependencyDefinition, builder.Configuration);
-        return builder;
-    }
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="builder"></param>
-    /// <typeparam name="TDependency"></typeparam>
-    /// <returns></returns>
-    public static IHostApplicationBuilder AddSekibanWithDependency<TDependency>(this IHostApplicationBuilder builder)
-        where TDependency : IDependencyDefinition, new()
-    {
-        builder.Services.AddSekibanWithDependency(new TDependency(), builder.Configuration);
-        return builder;
-    }
 
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="builder"></param>
-    /// <param name="dependencyDefinition"></param>
-    /// <param name="settings"></param>
-    /// <returns></returns>
-    public static IHostApplicationBuilder AddSekibanWithDependency(
-        this IHostApplicationBuilder builder,
-        IDependencyDefinition dependencyDefinition,
-        SekibanSettings settings)
-    {
-        builder.Services.AddSekibanWithDependency(dependencyDefinition, settings);
-        return builder;
-    }
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="services"></param>
-    /// <param name="configuration"></param>
-    /// <returns></returns>
-    public static IServiceCollection AddSekibanWithDependency<TDependency>(
-        this IServiceCollection services,
-        IConfiguration configuration) where TDependency : IDependencyDefinition, new()
-    {
-        var dependencyDefinition = new TDependency();
-        var settings = SekibanSettings.FromConfiguration(configuration);
-        return AddSekibanCoreWithDependency(services, dependencyDefinition, settings);
-    }
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="services"></param>
-    /// <param name="dependencyDefinition"></param>
-    /// <param name="configuration"></param>
-    /// <returns></returns>
-    public static IServiceCollection AddSekibanWithDependency(
-        this IServiceCollection services,
-        IDependencyDefinition dependencyDefinition,
-        IConfiguration configuration)
-    {
-        var settings = SekibanSettings.FromConfiguration(configuration);
-        return AddSekibanCoreWithDependency(services, dependencyDefinition, settings);
-    }
-
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="services"></param>
-    /// <param name="dependencyDefinition"></param>
-    /// <param name="settings"></param>
-    /// <returns></returns>
-    public static IServiceCollection AddSekibanWithDependency(
-        this IServiceCollection services,
-        IDependencyDefinition dependencyDefinition,
-        SekibanSettings settings) =>
-        AddSekibanCoreWithDependency(services, dependencyDefinition, settings);
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="services"></param>
-    /// <param name="dependencyDefinition"></param>
-    /// <param name="settings"></param>
-    /// <param name="sekibanDateProducer"></param>
-    /// <param name="multiProjectionType"></param>
-    /// <returns></returns>
-    public static IServiceCollection AddSekibanCoreWithDependency(
-        this IServiceCollection services,
-        IDependencyDefinition dependencyDefinition,
-        SekibanSettings settings,
-        ISekibanDateProducer? sekibanDateProducer = null,
-        SekibanCoreServiceExtensions.MultiProjectionType multiProjectionType
-            = SekibanCoreServiceExtensions.MultiProjectionType.MemoryCache)
-    {
-        Register(services, dependencyDefinition, settings, sekibanDateProducer, multiProjectionType);
-        return services;
-    }
-    /// <summary>
-    ///     Add Sekiban with Dependency
-    /// </summary>
-    /// <param name="services"></param>
-    /// <param name="dependencyDefinition"></param>
-    /// <param name="section"></param>
-    /// <returns></returns>
-    public static IServiceCollection AddSekibanWithDependencyWithConfigurationSection(
-        this IServiceCollection services,
-        IDependencyDefinition dependencyDefinition,
-        IConfigurationSection section)
-    {
-        var settings = SekibanSettings.FromConfigurationSection(section);
-        return AddSekibanCoreWithDependency(services, dependencyDefinition, settings);
-    }
     /// <summary>
     ///     Get Dependencies
     /// </summary>
@@ -147,40 +30,6 @@ public static class SekibanEventSourcingDependency
             typeof(ReportVersionToSnapshotManger.Handler));
     }
 
-    public static void Register(
-        IServiceCollection services,
-        IDependencyDefinition dependencyDefinition,
-        SekibanSettings settings,
-        ISekibanDateProducer? sekibanDateProducer = null,
-        SekibanCoreServiceExtensions.MultiProjectionType multiProjectionType
-            = SekibanCoreServiceExtensions.MultiProjectionType.MemoryCache)
-    {
-        // MediatR
-        services.AddMediatR(
-            new MediatRServiceConfiguration().RegisterServicesFromAssemblies(
-                Assembly.GetExecutingAssembly(),
-                GetAssembly()));
-        // Sekiban Event Sourcing
-        services.AddSekibanCore(settings, sekibanDateProducer ?? new SekibanDateProducer(), multiProjectionType);
-        services.AddSekibanConstantUser();
-        services.AddSingleton(settings);
-        services.AddTransient<IAggregateSettings, ContextAggregateSettings>();
-        // run Define() before using.
-        dependencyDefinition.Define();
-        // Each Domain contexts
-        services.AddSingleton(dependencyDefinition.GetSekibanDependencyOptions().RegisteredEventTypes);
-        services.AddSingleton(dependencyDefinition.GetSekibanDependencyOptions().SekibanAggregateTypes);
-        services.AddTransient(dependencyDefinition.GetSekibanDependencyOptions().TransientDependencies);
-        services.AddTransient(GetDependencies());
-
-        services.AddQueriesFromDependencyDefinition(dependencyDefinition);
-
-        foreach (var action in dependencyDefinition.GetServiceActions())
-        {
-            action(services);
-        }
-
-    }
     /// <summary>
     ///     Add Sekiban with Dependency for InMemory Test
     /// </summary>
