+++ src/Sekiban.Infrastructure.Cosmos/Sekiban.Infrastructure.Cosmos.csproj
@@ -37,7 +37,6 @@
     <ItemGroup>
         <ProjectReference Include="..\Sekiban.Core.DotNet\Sekiban.Core.DotNet.csproj"/>
         <ProjectReference Include="..\Sekiban.Infrastructure.Azure.Storage.Blobs\Sekiban.Infrastructure.Azure.Storage.Blobs.csproj"/>
-        <ProjectReference Include="..\Sekiban.Testing\Sekiban.Testing.csproj"/>
     </ItemGroup>
