+++ tests/Sekiban.Test.CosmosDb/CosmosSekibanServiceProviderGenerator.cs
@@ -1,10 +1,13 @@
+using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Sekiban.Core.Dependency;
 using Sekiban.Core.Shared;
 using Sekiban.Infrastructure.Azure.Storage.Blobs;
+using Sekiban.Infrastructure.Cosmos;
 using Sekiban.Testing.Story;
-namespace Sekiban.Infrastructure.Cosmos;
+
+namespace Sekiban.Test.CosmosDb;
 
 /// <summary>
 ///     Add Cosmos services
